requirements.txt
altair
anyio
beautifulsoup4
chromadb
fastapi
fastapi-cors
html5lib
langchain
langchain-chroma
langchain-community
langchain-core
langchain-experimental
langchain-groq
langchain-ollama
langchainhub
langgraph
langsmith
lxml
matplotlib
nest-asyncio
numpy
pandas
pluggy
psycopg2-binary
pymongo
pytest
python-dotenv
requests
scikit-learn
seaborn
SQLAlchemy
streamlit
tabulate
tavily-python
tiktoken
uvicorn
yfinance
pydantic
